Move modules to be more consistent with the rest
