+++ axum-extra/src/extract/user_lang/config.rs
@@ -1,5 +1,6 @@
-use crate::user_lang::{UserLanguage, UserLanguageSource};
 use std::sync::Arc;
+
+use crate::extract::user_lang::{UserLanguage, UserLanguageSource};
 
 /// TBD
 #[derive(Debug, Clone)]
