+++ axum-extra/src/extract/user_lang/mod.rs
@@ -1,4 +1,4 @@
-//! TBD
+//! User language extractor.
 
 mod config;
 mod lang;
