+++ axum-extra/src/extract/user_lang/sources/header.rs
@@ -1,6 +1,7 @@
-use crate::user_lang::UserLanguageSource;
 use axum::async_trait;
 use std::cmp::Ordering;
+
+use crate::extract::user_lang::UserLanguageSource;
 
 /// TBD
 #[derive(Debug, Clone)]
