+++ axum-extra/src/extract/user_lang/sources/query.rs
@@ -1,6 +1,7 @@
-use crate::user_lang::UserLanguageSource;
 use axum::{async_trait, extract::Query, RequestPartsExt};
 use std::collections::HashMap;
+
+use crate::extract::user_lang::UserLanguageSource;
 
 /// TBD
 #[derive(Debug, Clone)]
